refactor: apply resource build on `web/main.go`
